+++ ui/src/components/index.ts
@@ -5,6 +5,7 @@ export { Overlay } from './Overlay';
 export { Timer } from './Timer';
 export { QrPanel } from './QrPanel';
 export { PlayerRoster, type RosterPlayer } from './PlayerRoster';
+export { RoundScoreboard, type ScoreRow } from './RoundScoreboard';
 export { Avatar } from './Avatar';
 export { AvatarEditor } from './AvatarEditor';
 export { TextAnswerInput } from './TextAnswerInput';